backend/api/requirements.txt
Flask==3.0.3
google-cloud-storage==2.19.0
